workers/requirements.txt
celery==5.2.7
httpx==0.23.0
Jinja2==3.1.2
pika==1.3.0
pydantic==1.10.2
python-dotenv==0.20.0
requests==2.25.0
